Crud basico completo

+++ reportes-incidentes/src/main/java/com/javaintermedio/api/reportesincidentes/repository/EmpleadoRecursosHumanosRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmpleadoRecursosHumanosRepository extends JpaRepository<EmpleadoRecursosHumanos, Integer>{
+public interface EmpleadoRecursosHumanosRepository extends JpaRepository<EmpleadoRecursosHumanos, Long>{
     
 }
